P3_Percobaan 3_Pertanyaan 4_07

+++ src/Jobsheet_3/ArrayObjects/src/Segitiga/Segitiga.java
@@ -11,7 +11,7 @@ public class Segitiga {
     public double hitungLuas() {
         return 0.5 * alas * tinggi;
     }
-    public int hitungKeliling() {
+    public double hitungKeliling() {
         int sisiMiring = (int) Math.sqrt(Math.pow(alas,2) + Math.pow(tinggi,2));
         return alas + tinggi + sisiMiring;
     }
